third_party/WebKit/LayoutTests/fast/css/fontfaceset-check-platform-fonts.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script>

test(() => {
  assert_true(document.fonts.check('10px Arial'));
  assert_false(document.fonts.check('10px Nonexistent'));
  assert_true(document.fonts.check('10px sans-serif'));
  assert_true(document.fonts.check('10px Nonexistent, monospace'));
  assert_false(document.fonts.check('10px Nonexistent1, Nonexistent2'));
}, 'Tests FontFaceSet#check() returns true for platform fonts');

</script>
